[hotfix] move HiveTypeUtil to util package 'org.apache.flink.table.catalog.hive.util'
